Midterm for Network Programming push
This push includes programs assigned as a midterm project. One of the programs reads a file given as command line arguments and returns all IP addresses associated with a host, and the other program splits an array into three to find the max number within the three arrays, and returns the max value of those three.